done: todo

+++ todo_command.rb
@@ -1,15 +1,3 @@
-#listの書き込み、削除、Done、確認
-#
-# todo add "jsndcjn"
-#
-# todo del タスク番号
-#
-# todo check
-#
-# todo done タスク番号
-#
-# todo done list
-
 # タスクの追加
 def add(new_task,list)
   File.open(list, "a") do |f|
@@ -18,17 +6,38 @@ def add(new_task,list)
 end
 
 # タスクの削除
-def delete(tsak_num,todo_list)
+def delete(task_num,todo_list)
   delete_task = ""
+  todo_array = []
 
-  File.open(todo_list, "w+") do |f|
-    todo_array = f.readlines
+  File.open(todo_list, "r") do |file|
+
+    todo_array = file.readlines
+
     delete_task = todo_array[task_num-1]
+
+    if delete_task.nil?
+      STDERR.print "The task does not exist\n"
+      file.close
+      exit
+    end
+
     todo_array.delete_at(task_num-1)
-    todo_array.each{|todo| f.print(todo)}
+  end
+
+  File.open(todo_list, "w") do |file|
+    todo_array.each{|todo| file.print(todo)}
   end
 
   return delete_task
+end
+
+# todoからdoneへのタスクの移動
+def done(task_num,todo_list,done_list)
+  done_task = delete(task_num,todo_list)
+  add(done_task,done_list)
+
+  return done_task
 end
 
 # リストの確認
@@ -45,15 +54,14 @@ def check(comm,list)
   print list_name
   puts "<><><><><><><>"
 
+  i = 1
   File.open(list, "r") do |f|
     tasks = f.readlines
-    tasks.each{|task| print(task)}
+    tasks.each do |task|
+      print("   #{i}  #{task}")
+      i += 1
+    end
   end
 
   puts "<><><><><><><><><><><><><><><><><><><><><>"
-end
-
-def done(task_num,todo_list,done_list)
-  done_task = delete(task_num,todo_list)
-  add(done_task,done_list)
 end
